googleapis_auth: refactor the libraries to follow conventions
Prepare to release v1.1.0
