+++ googleapis_auth/lib/auth_browser.dart
@@ -2,19 +2,22 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library googleapis_auth.auth_browser;
-
 import 'dart:async';
 
 import 'package:http/browser_client.dart';
 import 'package:http/http.dart';
 
-import 'auth.dart';
+import 'src/access_credentials.dart';
+import 'src/auth_client.dart';
+import 'src/auth_functions.dart';
 import 'src/auth_http_utils.dart';
+import 'src/client_id.dart';
 import 'src/http_client_base.dart';
 import 'src/oauth2_flows/implicit.dart';
+import 'src/response_type.dart';
+import 'src/service_account_credentials.dart';
 
-export 'auth.dart';
+export 'googleapis_auth.dart';
 
 /// Obtains a HTTP client which uses the given [apiKey] for making HTTP
 /// requests.
