+++ googleapis_auth/lib/src/auth_http_utils.dart
@@ -2,13 +2,15 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library googleapis_auth;
-
 import 'dart:async';
 
 import 'package:http/http.dart';
 
-import '../auth.dart';
+import 'access_credentials.dart';
+import 'auth_client.dart';
+import 'auth_functions.dart';
+import 'client_id.dart';
+import 'exceptions.dart';
 import 'http_client_base.dart';
 
 /// Will close the underlying `http.Client` depending on a constructor argument.
