+++ googleapis_auth/lib/src/client_id.dart
@@ -1,0 +1,26 @@
+// Copyright (c) 2021, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+/// Represents the client application's credentials.
+class ClientId {
+  /// The identifier used to identify this application to the server.
+  final String identifier;
+
+  /// The client secret used to identify this application to the server.
+  final String? secret;
+
+  ClientId(this.identifier, this.secret);
+
+  ClientId.serviceAccount(this.identifier) : secret = null;
+
+  factory ClientId.fromJson(Map<String, dynamic> json) => ClientId(
+        json['identifier'] as String,
+        json['secret'] as String?,
+      );
+
+  Map<String, dynamic> toJson() => {
+        'identifier': identifier,
+        if (secret != null) 'secret': secret,
+      };
+}
