+++ googleapis_auth/lib/src/crypto/asn1.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library appengine_auth.asn;
-
 import 'dart:typed_data';
 
 import 'rsa.dart';
