+++ googleapis_auth/lib/src/crypto/pem.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library googleapis_auth.pem;
-
 import 'dart:convert';
 import 'dart:typed_data';
