+++ googleapis_auth/lib/src/crypto/rsa_sign.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library googleapis_auth.rsa_sign;
-
 import 'dart:typed_data';
 
 import 'package:crypto/crypto.dart';
