+++ googleapis_auth/lib/src/http_client_base.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library googleapis_auth.http_client_base;
-
 import 'dart:async';
 
 import 'package:http/http.dart';
