+++ googleapis_auth/lib/src/oauth2_flows/jwt.dart
@@ -2,14 +2,13 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library jwt_token_generator;
-
 import 'dart:async';
 import 'dart:convert';
 
 import 'package:http/http.dart' as http;
 
-import '../../auth.dart';
+import '../access_credentials.dart';
+import '../access_token.dart';
 import '../crypto/rsa.dart';
 import '../crypto/rsa_sign.dart';
 import '../http_client_base.dart';
