+++ googleapis_auth/lib/src/oauth2_flows/metadata_server.dart
@@ -1,8 +1,9 @@
 // Copyright (c) 2014, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
-
-library googleapis_auth.metadata_server_flow;
+// Copyright (c) 2021, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
 
 import 'dart:async';
 import 'dart:convert';
@@ -10,7 +11,8 @@ import 'dart:io';
 
 import 'package:http/http.dart' as http;
 
-import '../../auth.dart';
+import '../access_credentials.dart';
+import '../access_token.dart';
 import '../utils.dart';
 
 /// Obtains access credentials form the metadata server.
