+++ googleapis_auth/test/crypto/asn1_test.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library googleapis_auth.test.asn1_test;
-
 import 'dart:typed_data';
 
 import 'package:googleapis_auth/src/crypto/asn1.dart';
