+++ googleapis_auth/test/crypto/rsa_sign_test.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library googleapis_auth.test.crypto_rsa_sign_test;
-
 import 'dart:convert';
 
 import 'package:googleapis_auth/src/crypto/rsa_sign.dart';
