+++ googleapis_auth/test/crypto/rsa_test.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library googleapis_auth.test.rsa_test;
-
 import 'package:googleapis_auth/src/crypto/rsa.dart';
 import 'package:test/test.dart';
