+++ googleapis_auth/test/http_client_base_test.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library googleapis_auth.http_client_base_test;
-
 import 'dart:async';
 
 import 'package:googleapis_auth/src/auth_http_utils.dart';
