+++ googleapis_auth/test/oauth2_flows/auth_code_test.dart
@@ -2,13 +2,11 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library googleapis_auth.auth_code_test;
-
 import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:googleapis_auth/auth.dart';
+import 'package:googleapis_auth/googleapis_auth.dart';
 import 'package:googleapis_auth/src/oauth2_flows/auth_code.dart';
 import 'package:http/http.dart';
 import 'package:test/test.dart';
