+++ googleapis_auth/test/oauth2_flows/jwt_test.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library googleapis_auth.jwt_test;
-
 import 'dart:async';
 import 'dart:convert';
