+++ googleapis_auth/test/oauth2_test.dart
@@ -2,12 +2,10 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library googleapis_auth.oauth2_test;
-
 import 'dart:async';
 import 'dart:convert';
 
-import 'package:googleapis_auth/auth.dart';
+import 'package:googleapis_auth/googleapis_auth.dart';
 import 'package:googleapis_auth/src/http_client_base.dart';
 import 'package:googleapis_auth/src/utils.dart';
 import 'package:http/http.dart';
